js fixed docker issues(sucks) and created new image branch and pushing everything(its late this sucks)
