+++ client/Dockerfile
@@ -1,19 +1,24 @@
-FROM node:20-slim
+# First stage: Build
+FROM node:20 AS builder
 
-# Set working directory
 WORKDIR /app
 
-# Copy only what's needed
 COPY package*.json ./
-
-# Install deps first (Docker layer caching!)
 RUN npm install
 
-# Copy remaining files
 COPY . .
+RUN npm run build
 
-# Expose port
+# Second stage: Serve only production
+FROM node:20-slim
+
+WORKDIR /app
+
+COPY package*.json ./
+RUN npm install --omit=dev --ignore-scripts
+
+COPY --from=builder /app/.next ./.next
+COPY --from=builder /app/public ./public
+
 EXPOSE 3000
-
-# Run app
-CMD ["npm", "run", "dev"]
+CMD ["npm", "start"]
